requirements.txt
Flask
celery[google]
python-dotenv